Make explorer's available views widget a pure React component
This new react component ReactAvailableViewsWidget will be part of the
packages/react-components. It can be used in other web applications
independently from Theia.

It listens to the AVAILABLE_OUTPUTS_CHANGED Signal to update the widget.
